Fixed even more bugs in install.sh
I have extensively tested it on a debian vm, so it should work without
any problems in this enviroment. I still have to test it on other
distros, I'm thinking Ubuntu, Arch and Fedora. This way the majority of
setups will be covered.
Also, I still have to manage the output of the script and add a few
options to specify verbosity.